+++ app/models/js_ignore.rb
