+++ spec/models/config/jshint_spec.rb
@@ -1,5 +1,4 @@
-require "spec_helper"
-require "lib/js_ignore"
+require "app/models/js_ignore"
 require "app/models/config/base"
 require "app/models/config/jshint"
 require "app/models/config/parser"
