+++ spec/models/github_api_spec.rb
@@ -1,7 +1,6 @@
-require "spec_helper"
-require "attr_extras"
-require "lib/github_api"
+require "octokit"
 require "json"
+require "app/models/github_api"
 
 describe GithubApi do
   before { stub_const("Hound::GITHUB_TOKEN", "token") }
